% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findParent.R
\name{findParent}
\alias{findParent}
\title{Find children and parents}
\usage{
findParent(Sentence, Path = "amod")
}
\arguments{
\item{Sentence}{a record (row) in the GeoDeepDive NLP output}

\item{Path}{a stanford core nlp pose code}
}
\value{
A character matrix
}
\description{
Returns a tuple of a child and a parent defined by the StanfordCoreNLP.
}
\details{

}
\examples{

# TBDDDDD

}
\author{
Andrew A. Zaffos & Erika T. Ito
}
